+++ tests/Gedmo/Mapping/TranslatableMappingTest.php
@@ -11,6 +11,8 @@ declare(strict_types=1);
 
 namespace Gedmo\Tests\Mapping;
 
+use Doctrine\Common\EventManager;
+use Doctrine\ORM\EntityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\Mapping\Driver\YamlDriver;
 use Doctrine\Persistence\Mapping\Driver\MappingDriverChain;
@@ -18,14 +20,13 @@ use Gedmo\Mapping\ExtensionMetadataFactory;
 use Gedmo\Tests\Mapping\Fixture\Yaml\User;
 use Gedmo\Tests\Translatable\Fixture\PersonTranslation;
 use Gedmo\Translatable\TranslatableListener;
-use Symfony\Component\Cache\Adapter\ArrayAdapter;
 
 /**
  * These are mapping tests for translatable behavior
  *
  * @author Gediminas Morkevicius <gediminas.morkevicius@gmail.com>
  */
-final class TranslatableMappingTest extends \PHPUnit\Framework\TestCase
+final class TranslatableMappingTest extends ORMMappingTestCase
 {
     public const TEST_YAML_ENTITY_CLASS = User::class;
 
@@ -41,11 +42,9 @@ final class TranslatableMappingTest extends \PHPUnit\Framework\TestCase
 
     protected function setUp(): void
     {
-        $config = new \Doctrine\ORM\Configuration();
-        $config->setMetadataCache(new ArrayAdapter());
-        $config->setQueryCache(new ArrayAdapter());
-        $config->setProxyDir(TESTS_TEMP_DIR);
-        $config->setProxyNamespace('Gedmo\Mapping\Proxy');
+        parent::setUp();
+
+        $config = $this->getBasicConfiguration();
         $chainDriverImpl = new MappingDriverChain();
         $chainDriverImpl->addDriver(
             new YamlDriver([__DIR__.'/Driver/Yaml']),
@@ -58,13 +57,12 @@ final class TranslatableMappingTest extends \PHPUnit\Framework\TestCase
             'memory' => true,
         ];
 
-        //$config->setSQLLogger(new \Doctrine\DBAL\Logging\EchoSQLLogger());
-
-        $evm = new \Doctrine\Common\EventManager();
+        $evm = new EventManager();
         $this->translatableListener = new TranslatableListener();
+        $this->translatableListener->setCacheItemPool($this->cache);
         $this->translatableListener->setTranslatableLocale('en_us');
         $evm->addEventSubscriber($this->translatableListener);
-        $this->em = \Doctrine\ORM\EntityManager::create($conn, $config, $evm);
+        $this->em = EntityManager::create($conn, $config, $evm);
     }
 
     public function testYamlMapping(): void
@@ -74,7 +72,7 @@ final class TranslatableMappingTest extends \PHPUnit\Framework\TestCase
             self::TEST_YAML_ENTITY_CLASS,
             'Gedmo\Translatable'
         );
-        $config = $this->em->getMetadataFactory()->getCacheDriver()->fetch($cacheId);
+        $config = $this->cache->getItem($cacheId)->get();
         static::assertArrayHasKey('translationClass', $config);
         static::assertSame(PersonTranslation::class, $config['translationClass']);
         static::assertArrayHasKey('fields', $config);